Had to update the cars_id field to id in order to get my endpoints working. Will be testing the additional post, update, delete endpoints tomorrow.
