+++ package.json
@@ -14,6 +14,8 @@
   },
   "dependencies": {
     "dotenv": "^8.2.0",
+    "express": "^4.17.1",
+    "helmet": "^3.21.2",
     "jest": "^24.9.0",
     "knex": "^0.20.4",
     "sqlite3": "^4.1.1"
