+++ routes/carsRouter.js
@@ -19,6 +19,8 @@ router.get('/:id', async (req, res, next) => {
         const cars = await db('cars')
                             .where({ id: req.params.id})
                             .first()
+        res
+            .json(cars)
     }
     catch (error) {
         next(error)
@@ -41,3 +43,5 @@ router.post('/', async (req, res, next) => {
         next(error)
     }
 })
+
+module.exports = router
