+++ server.js
@@ -1,24 +1,24 @@
-const express = require('express')
-const carsRouter = require('./routes/carsRouter')
-// const db = require('./<whatever my db file is here');
+// const express = require('express')
+// const carsRouter = require('./routes/carsRouter')
+// // const db = require('./<whatever my db file is here');
 
-const server = express();
+// const server = express();
 
-server.use(express.json());
-server.use('/api/cars', carsRouter)
+// server.use(express.json());
+// server.use('/api/cars', carsRouter)
 
 
-server.use((req, res) => {
-    res
-        .status(404)
-        .json({ message: "Route was not found." })
-    })
+// server.use((req, res) => {
+//     res
+//         .status(404)
+//         .json({ message: "Route was not found." })
+//     })
 
-server.use((err, req, res, next) => {
-    console.log(err)
-    res
-        .status(500)
-        .json({ message: "Something went wrong." })
-})
+// server.use((err, req, res, next) => {
+//     console.log(err)
+//     res
+//         .status(500)
+//         .json({ message: "Something went wrong." })
+// })
 
-module.exports = server;
+// module.exports = server;
